Issue #3056204 by Schoonzie: Allow Secret Name to be configurable and support key/value pair secrets

+++ src/Plugin/KeyProvider/AwsSecretsManagerKeyProvider.php
@@ -5,9 +5,11 @@ namespace Drupal\aws_secrets_manager\Plugin\KeyProvider;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Psr\Log\LoggerInterface;
 use Aws\SecretsManager\SecretsManagerClient;
+use Drupal\Core\Form\FormStateInterface;
 use Drupal\key\KeyInterface;
 use Drupal\key\Plugin\KeyProviderBase;
 use Drupal\key\Plugin\KeyProviderSettableValueInterface;
+use Drupal\key\Plugin\KeyPluginFormInterface;
 
 /**
  * Adds a key provider that allows a key to be stored in AWS Secrets Manager.
@@ -19,11 +21,11 @@ use Drupal\key\Plugin\KeyProviderSettableValueInterface;
  *   storage_method = "aws_secrets_manager",
  *   key_value = {
  *     "accepted" = TRUE,
- *     "required" = TRUE
+ *     "required" = FALSE
  *   }
  * )
  */
-class AwsSecretsManagerKeyProvider extends KeyProviderBase implements KeyProviderSettableValueInterface {
+class AwsSecretsManagerKeyProvider extends KeyProviderBase implements KeyProviderSettableValueInterface, KeyPluginFormInterface {
 
   /**
    * The settings.
@@ -97,14 +99,63 @@ class AwsSecretsManagerKeyProvider extends KeyProviderBase implements KeyProvide
    * {@inheritdoc}
    */
   public function defaultConfiguration() {
-    return [];
+    return [
+      'secret_name' => '',
+      'property_name' => '',
+      'read_only' => TRUE,
+    ];
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function buildConfigurationForm(array $form, FormStateInterface $form_state) {
+    $form['secret_name'] = [
+      '#type' => 'textfield',
+      '#title' => $this->t('Secret Name'),
+      '#description' => $this->t('The name of the secret in AWS Secrets Manager.'),
+      '#required' => TRUE,
+      '#default_value' => $this->getConfiguration()['secret_name'] ?? '',
+    ];
+
+    $form['property_name'] = [
+      '#type' => 'textfield',
+      '#title' => $this->t('Property Name'),
+      '#description' => $this->t('If the secret is a JSON object, what is the property name that contains the value.'),
+      '#required' => FALSE,
+      '#default_value' => $this->getConfiguration()['property_name'] ?? '',
+    ];
+
+    $form['read_only'] = [
+      '#title' => $this->t('Read only'),
+      '#description' => $this->t('Whether or not we should try to update this value in AWS Secrets Manager.'),
+      '#type' => 'checkbox',
+      '#default_value' => $this->getConfiguration()['read_only'] ?? FALSE,
+    ];
+
+    return $form;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function validateConfigurationForm(array &$form, FormStateInterface $form_state) {
+    $key_provider_settings = $form_state->getValues();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function submitConfigurationForm(array &$form, FormStateInterface $form_state) {
+    $this->setConfiguration($form_state->getValues());
   }
 
   /**
    * {@inheritdoc}
    */
   public function getKeyValue(KeyInterface $key) {
-    $name = $key->id();
+    $name = $this->getConfiguration()['secret_name'] ?? $key->id();
+    $property_name = $this->getConfiguration()['property_name'] ?? '';
 
     try {
       $response = $this->client->getSecretValue([
@@ -112,6 +163,16 @@ class AwsSecretsManagerKeyProvider extends KeyProviderBase implements KeyProvide
       ]);
 
       if ($value = $response->get('SecretString')) {
+        if (!empty($property_name)) {
+          $value = json_decode($value);
+          if (!isset($value->{$property_name})) {
+            $this->logger->error(sprintf("unable to find property %p in secret %s", $property_name, $name));
+            return '';
+          }
+          else {
+            return $value->{$property_name};
+          }
+        }
         return $value;
       }
     }
@@ -126,8 +187,21 @@ class AwsSecretsManagerKeyProvider extends KeyProviderBase implements KeyProvide
    * {@inheritdoc}
    */
   public function setKeyValue(KeyInterface $key, $key_value) {
-    $name = $key->id();
+    $read_only = $this->getConfiguration()['read_only'] ?? FALSE;
+    $name = $this->getConfiguration()['secret_name'] ?? $key->id();
+
+    if ($read_only) {
+      $this->logger->error(sprintf("can't update read only secret %s", $name));
+      return FALSE;
+    }
+
     $label = $key->label();
+
+    $property_name = $this->getConfiguration()['property_name'] ?? '';
+
+    if (!empty($property_name)) {
+      $key_value = [$property_name => $key_value];
+    }
 
     try {
       $response = $this->client->createSecret([
@@ -148,7 +222,13 @@ class AwsSecretsManagerKeyProvider extends KeyProviderBase implements KeyProvide
    * {@inheritdoc}
    */
   public function deleteKeyValue(KeyInterface $key) {
-    $name = $key->id();
+    $read_only = $this->getConfiguration()['read_only'] ?? FALSE;
+    $name = $this->getConfiguration()['secret_name'] ?? $key->id();
+
+    if ($read_only) {
+      $this->logger->error(sprintf("can't delete read only secret %s", $name));
+      return FALSE;
+    }
 
     try {
       $response = $this->client->deleteSecret([
